Rename dump/restore to export/import
