+++ features/Role.feature
@@ -1,18 +1,18 @@
 Feature: Role Commands
 
-  # role:dump & role:restore
+  # role:export & role:import
 
   @security
-  Scenario: Dump and restore roles
+  Scenario: Export and import roles
     Given I create a role "teacher" with the following API rights:
       | document | { "actions": { "create": true } } |
     And I create a role "student" with the following API rights:
       | document | { "actions": { "update": true } } |
-    When I run the command "role:dump" with:
+    When I run the command "role:export" with:
       | flag | --path | ./dump |
     And I delete the role "teacher"
     And I delete the role "student"
-    When I run the command "role:restore" with:
+    When I run the command "role:import" with:
       | arg | ./dump/roles.json |  |
     Then The role "teacher" should match:
       | document | { "actions": { "create": true } } |
