+++ src/commands/collection/export.ts
@@ -3,16 +3,15 @@ import { Kommand } from '../../common'
 import { kuzzleFlags, KuzzleSDK } from '../../support/kuzzle'
 import { dumpCollectionData, dumpCollectionMappings } from '../../support/dump-collection'
 import * as fs from 'fs'
-import cli from 'cli-ux'
 import chalk from 'chalk'
 
-export default class IndexDump extends Kommand {
-  static description = 'Dump an entire index content (JSONL format)'
+export default class CollectionExport extends Kommand {
+  static description = 'Exports an collection (JSONL format)'
 
   static flags = {
     help: flags.help({}),
     path: flags.string({
-      description: 'Dump directory (default: index name)',
+      description: 'Dump root directory (default: index name)',
     }),
     'batch-size': flags.string({
       description: 'Maximum batch size (see limits.documentsFetchCount config)',
@@ -23,6 +22,7 @@ export default class IndexDump extends Kommand {
 
   static args = [
     { name: 'index', description: 'Index name', required: true },
+    { name: 'collection', description: 'Collection name', required: true },
   ]
 
   async run() {
@@ -37,38 +37,30 @@ export default class IndexDump extends Kommand {
   async runSafe() {
     this.printCommand()
 
-    const { args, flags: userFlags } = this.parse(IndexDump)
+    const { args, flags: userFlags } = this.parse(CollectionExport)
 
     const path = userFlags.path || args.index
 
     const sdk = new KuzzleSDK(userFlags)
     await sdk.init(this.log)
 
-    this.log(chalk.green(`Dumping index "${args.index}" in ${path}/ ...`))
+    this.log(chalk.green(`Dumping collection "${args.index}:${args.collection}" in ${path}/ ...`))
 
     fs.mkdirSync(path, { recursive: true })
 
-    const { collections } = await sdk.collection.list(args.index)
+    await dumpCollectionMappings(
+      sdk,
+      args.index,
+      args.collection,
+      path)
 
-    for (const collection of collections) {
-      if (collection.type !== 'realtime') {
-        await dumpCollectionMappings(
-          sdk,
-          args.index,
-          collection.name,
-          path)
+    await dumpCollectionData(
+      sdk,
+      args.index,
+      args.collection,
+      Number(userFlags['batch-size']),
+      path)
 
-        await dumpCollectionData(
-          sdk,
-          args.index,
-          collection.name,
-          Number(userFlags['batch-size']),
-          path)
-
-        cli.action.stop()
-      }
-    }
-
-    this.log(chalk.green(`[✔] Index ${args.index} dumped`))
+    this.log(chalk.green(`[✔] Collection ${args.index}:${args.collection} dumped`))
   }
 }
