+++ src/commands/profile/export.ts
@@ -4,20 +4,20 @@ import { kuzzleFlags, KuzzleSDK } from '../../support/kuzzle'
 import * as fs from 'fs'
 import chalk from 'chalk'
 
-export default class RoleDump extends Kommand {
+export default class ProfileExport extends Kommand {
   private batchSize?: string;
 
   private path?: string;
 
   private sdk?: any;
 
-  static description = 'Dumps Kuzzle roles'
+  static description = 'Exports profiles'
 
   static flags = {
     help: flags.help({}),
     path: flags.string({
       description: 'Dump directory',
-      default: 'roles'
+      default: 'profiles'
     }),
     'batch-size': flags.string({
       description: 'Maximum batch size (see limits.documentsFetchCount config)',
@@ -38,45 +38,45 @@ export default class RoleDump extends Kommand {
   async runSafe() {
     this.printCommand()
 
-    const { flags: userFlags } = this.parse(RoleDump)
+    const { flags: userFlags } = this.parse(ProfileExport)
 
     this.path = userFlags.path
     this.batchSize = userFlags['batch-size']
 
-    const filename = `${this.path}/roles.json`
+    const filename = `${this.path}/profiles.json`
 
     this.sdk = new KuzzleSDK(userFlags)
     await this.sdk.init(this.log)
 
-    this.log(`Dumping roles in ${filename} ...`)
+    this.log(`Dumping securities in ${filename} ...`)
 
     fs.mkdirSync(this.path, { recursive: true })
 
-    const roles = await this._dumpRoles()
+    const profiles = await this._dumpProfiles()
 
-    fs.writeFileSync(filename, JSON.stringify(roles, null, 2))
+    fs.writeFileSync(filename, JSON.stringify(profiles, null, 2))
 
-    this.log(chalk.green(`[✔] ${Object.keys(roles).length} roles dumped`))
+    this.log(chalk.green(`[✔] ${Object.keys(profiles).length} profiles dumped`))
   }
 
-  async _dumpRoles() {
+  async _dumpProfiles() {
     const options = {
       scroll: '10m',
       size: this.batchSize
     }
 
-    let result = await this.sdk.security.searchRoles({}, options)
+    let result = await this.sdk.security.searchProfiles({}, options)
 
-    const roles: any = {}
+    const profiles: any = {}
 
     while (result) {
       result.hits.forEach((hit: any) => {
-        roles[hit._id] = { controllers: hit.controllers }
+        profiles[hit._id] = { policies: hit.policies }
       })
 
       result = await result.next()
     }
 
-    return roles
+    return profiles
   }
 }
